requirements.txt
autopep8==1.5.2
boto3==1.14.43
botocore==1.17.43
Django==2.1.1
django-cors-headers==3.2.1
django-environ==0.4.5
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
environ==1.0
jmespath==0.10.0
Pillow==7.1.2
psycopg2==2.8.5
pycodestyle==2.6.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.11.0
urllib3==1.25.10
